+++ W040422MR - 2022.11.21/real-app-front/changelog.md
@@ -1,3 +1,8 @@
+## create common Input component
+
+- create Input component
+- 
+
 ## create signup page
 
 - create sign up component page